Make ghosts independant
As we talked about, there is now a class per ghost,
(no base class).

This in turns introduce a few function template.

One big down side is that whe have to use *this.

I also factored out ghost animation in a separate class to
try to tame code duplication a bit.